+++ database/migrations/2024_07_26_202628_create_dimentions_table.php
@@ -14,9 +14,9 @@ return new class extends Migration
         Schema::create('dimentions', function (Blueprint $table) {
             $table->id();
             $table->foreignUuid('product_id')->constrained()->onDelete('cascade');
-            $table->integer('length')->unsigned()->nullable()->default(0);
-            $table->integer('width')->unsigned()->nullable()->default(0);
-            $table->integer('height')->unsigned()->nullable()->default(0);
+            $table->decimal('length')->unsigned()->nullable()->default(0);
+            $table->decimal('width')->unsigned()->nullable()->default(0);
+            $table->decimal('height')->unsigned()->nullable()->default(0);
             $table->timestamps();
         });
     }
